bitwindow: style send page
