+++ clients/bitwindow/lib/pages/sidechain_proposal_page.dart
@@ -1,6 +1,5 @@
 import 'package:auto_route/annotations.dart';
 import 'package:bitwindow/servers/api.dart';
-import 'package:bitwindow/widgets/qt_button.dart';
 import 'package:bitwindow/widgets/qt_container.dart';
 import 'package:bitwindow/widgets/qt_icon_button.dart';
 import 'package:flutter/material.dart';
