+++ clients/bitwindow/lib/widgets/error_container.dart
@@ -1,4 +1,3 @@
-import 'package:bitwindow/widgets/qt_button.dart';
 import 'package:bitwindow/widgets/qt_container.dart';
 import 'package:flutter/material.dart';
 import 'package:sail_ui/sail_ui.dart';
