+++ clients/sail_ui/lib/widgets/inputs/dropdown_menu.dart
@@ -10,7 +10,6 @@ class SailDropdownButton<T> extends StatefulWidget {
       Icons.expand_more,
       size: 16,
     ),
-    this.width,
     this.large = false,
     this.enabled = true,
     super.key,
@@ -20,7 +19,6 @@ class SailDropdownButton<T> extends StatefulWidget {
   final ValueChanged<T> onChanged;
   final T value;
   final Widget? icon;
-  final double? width;
   final bool large;
   final bool enabled;
   @override
@@ -30,10 +28,11 @@ class SailDropdownButton<T> extends StatefulWidget {
 class _SailDropdownButtonState<T> extends State<SailDropdownButton<T>> {
   @override
   Widget build(BuildContext context) {
+    final theme = SailTheme.of(context);
+
     var items = widget.items
         .map(
           (e) => SailMenuItem(
-            height: widget.large ? 24 : 18,
             onSelected: () {
               widget.onChanged(e.value);
             },
@@ -45,53 +44,34 @@ class _SailDropdownButtonState<T> extends State<SailDropdownButton<T>> {
     var currentIndex = widget.items.indexWhere((element) => element.value == widget.value);
     Widget currentItem = widget.items[currentIndex];
 
-    var offsetY = 0.0;
-    for (var i = 0; i < currentIndex; i += 1) {
-      offsetY += items[i].height;
-    }
-
-    if (widget.width != null) {
-      currentItem = Expanded(
-        child: currentItem,
-      );
-    }
-
-    return SizedBox(
-      width: widget.width,
-      height: widget.large ? 32 : 24,
-      child: _Button(
-        enabled: widget.enabled,
-        large: widget.large,
-        padding: EdgeInsets.only(
-          left: 8,
-          right: widget.icon == null ? 8 : 4,
-        ),
-        onPressed: widget.enabled
-            ? () {
-                var bounds = getGlobalBoundsForContext(context);
-                showSailMenu(
-                  context: context,
-                  preferredAnchorPoint: Offset(
-                    bounds.left - (context.isWindows ? 1 : 9),
-                    bounds.top - offsetY - 3,
-                  ),
-                  menu: SailMenu(
-                    items: items,
-                    width: widget.width,
-                  ),
-                );
-              }
-            : null,
+    return QtButton(
+      size: ButtonSize.small,
+      onPressed: () {
+        var bounds = getGlobalBoundsForContext(context);
+        showSailMenu(
+          context: context,
+          preferredAnchorPoint: Offset(
+            bounds.left - (context.isWindows ? 1 : 9),
+            bounds.bottom,
+          ),
+          menu: SailMenu(
+            items: items,
+          ),
+        );
+      },
+      style: SailButtonStyle.secondary,
+      disabled: !widget.enabled,
+      child: Container(
+        padding: EdgeInsets.symmetric(horizontal: SailStyleValues.padding16),
         child: Row(
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           mainAxisSize: MainAxisSize.max,
           children: [
             currentItem,
-            if (widget.icon != null)
-              Padding(
-                padding: const EdgeInsets.only(left: 4),
-                child: widget.icon,
-              ),
+            SailSVG.fromAsset(
+              SailSVGAsset.iconDropdown,
+              color: theme.colors.textSecondary,
+            ),
           ],
         ),
       ),
@@ -114,68 +94,6 @@ class SailDropdownItem<T> extends StatelessWidget {
     return Align(
       alignment: Alignment.centerLeft,
       child: child,
-    );
-  }
-}
-
-class _Button extends StatelessWidget {
-  final VoidCallback? onPressed;
-  final Widget child;
-  final EdgeInsets padding;
-  final bool large;
-  final bool enabled;
-  const _Button({
-    required this.onPressed,
-    required this.child,
-    this.padding = const EdgeInsets.symmetric(
-      horizontal: 12,
-      vertical: 0,
-    ),
-    required this.large,
-    this.enabled = true,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    Color textColor;
-    final sailTheme = context.sailTheme;
-    if (onPressed != null && enabled) {
-      textColor = sailTheme.colors.text;
-    } else {
-      textColor = sailTheme.colors.textTertiary;
-    }
-
-    return SizedBox(
-      height: large ? 32 : 24,
-      child: MaterialButton(
-        // textTheme: textTheme,
-        mouseCursor: enabled ? WidgetStateMouseCursor.clickable : SystemMouseCursors.forbidden,
-        color: enabled ? sailTheme.colors.backgroundSecondary : sailTheme.colors.disabledBackground,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.all(Radius.circular(context.isWindows ? 3 : 4)),
-          side: BorderSide(
-            color: sailTheme.colors.formFieldBorder,
-            width: 1,
-          ),
-        ),
-        onPressed: onPressed,
-        elevation: 0,
-        hoverElevation: 0,
-        focusElevation: 0,
-        minWidth: 32,
-        padding: EdgeInsets.zero,
-        child: Container(
-          alignment: Alignment.center,
-          height: large ? 32 : 24,
-          padding: padding,
-          child: DefaultTextStyle(
-            style: TextStyle(
-              color: textColor,
-            ),
-            child: child,
-          ),
-        ),
-      ),
     );
   }
 }
